requirements.txt
customtkinter
sqlalchemy
psycopg2-binary
python-dotenv
